Changed datatypes.

+++ Section7/BankAccount.java
@@ -24,7 +24,7 @@ public class BankAccount {
         return balance;
     }
     //Creating a constructor to represent the act of making a deposit.
-    public float deposit()
+    public double deposit()
     {
         //Returning the amount to deposit.
         return deposit;
